Use official consul API client, register check only once
The Go client for the Consul API was moved to the main hashicorp consul
repo.

The health check for the boxrunner instance is registered only once.
Before this change, the health check was registered again and again in
case the session creation failed. For some reason the health seems to be
in critical state for a very short time after it is re-registered and
the session creation was always attempted during this time, causing an
infinite loop of unsuccessful session creation attempts.